requirements.txt
streamlit
opencv
pillow
numpy
matplotlib